Pb 10

+++ src/Pb_10/Main.java
@@ -3,7 +3,7 @@ import static Tools.Tools.*;
 
 public class Main {
     public static void main(String[] args) {
-        int sum = 2;
+        long sum = 2;
         for (int i = 3; i < 2000000; i += 1) {
             if (isPrime(i)) {
                 sum += i;
